refactor: split Icon into components/icon package — Icon, RevealIcon, ExpandingIcon (#7)
The merged Icon class multiplexed two interaction styles through
type="simple"|"pulse" (~50 params, most silently ignored per style).
Split into three siblings under components/icon/:

- icon.py: the classic Icon, restored to its public v1.0.3 form (only
  delta: import depth + a marked deprecation shim) so the next public
  release diff stays reviewable
- reveal.py: RevealIcon — the launch treatment (halo, breath-pulse
  dwell, expand, LaunchBlackoutOverlay) plus the launcher grid statics
- expanding.py: ExpandingIcon moved unchanged (components/expanding_icon
  kept as a re-export shim)

Icon(type="pulse") and the Icon.grid_*/layout_* statics still work via
deprecation shims that dispatch to RevealIcon (warning logged) so canopy
keeps running until its own migration PR; remove the shims before the
next public release. All existing import paths keep working; RavenApp's
home button now constructs RevealIcon directly.
