+++ __init__.py
@@ -37,9 +37,10 @@ from .helpers.utils_light import *
 _COMPONENTS = {
     "Button": "button",
     "Container": "container",
-    "ExpandingIcon": "expanding_icon",
+    "ExpandingIcon": "icon",
     "HorizontalContainer": "horizontal_container",
     "Icon": "icon",
+    "RevealIcon": "icon",
     "ScrollView": "scroll_view",
     "Spacer": "spacer",
     "TextBox": "text_box",
@@ -185,6 +186,7 @@ __all__ = [
     "HorizontalContainer",
     "Icon",
     "RavenApp",
+    "RevealIcon",
     "Routine",
     "RunApp",
     "ScrollView",
